Removing unnecessary compiler warning.
RXTX uses UTS_RELEASE to check the current kernel version (in order to compare to the one it was built under and give a warning if they don't match).  Unfortunately, the location of this define has moved in the kernel I have.  Since the warning isn't that useful anyway, I'm just commenting out the relevant check rather than fixing the include.
This is also what Debian did, it seems.